Update room.js
Previously, when assigning characters to players we were emitting in a real time loop iteratively. 

I've modified this code to first assign the roles in a locally scoped object first, once role assignment is complete, *then* we emit on the socket to each player, thus making the game more stable and less error prone.